✨Feat : Register, Modify 같이 사용하는 View 생성
