+++ OpenMarket/OpenMarket/Base.lproj/Main.storyboard
@@ -58,6 +58,21 @@
             </objects>
             <point key="canvasLocation" x="1042" y="726"/>
         </scene>
+        <!--Modify View Controller-->
+        <scene sceneID="RvG-gH-kMo">
+            <objects>
+                <viewController storyboardIdentifier="ModifyViewController" id="LVA-0u-qYa" customClass="ModifyViewController" customModule="OpenMarket" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="CPO-wq-Ipj">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <viewLayoutGuide key="safeArea" id="fRg-nX-GV9"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="Opj-vR-w2Z" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="125" y="721"/>
+        </scene>
     </scenes>
     <resources>
         <systemColor name="systemBackgroundColor">
